src/pyproject.toml
[project]
name = "customer-service-agent-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "memory-module",
    "python-dotenv",
    "aiohttp",
    "teams-ai",
    "botbuilder",
    "pydantic>=2.10.3",
]

[tool.uv.sources]
memory-module = { workspace = true }